Move the spot edit bahviour classes inside a utility class.
